restructure how the dash app is created
